Add files via upload
Choose a file name. 
The program would open that file and read in the Sudoku puzzle file that consist of 9 rows of 9 single digit numbers each, with zeroes representing the initially empty cells. Display the initial problem in a 9x9 grid. The program should then solve the puzzle by filling in the empty cells of the grid and the final solution.